+++ app.config.ts
@@ -19,7 +19,7 @@ const config = ({ config }: ConfigContext): ExpoConfig => {
   ...config,
   name: 'Expert Na Bíblia',
   slug: 'expert-na-biblia',
-  version: '1.8.0',
+  version: '1.9.0',
   orientation: 'portrait',
   icon: './assets/icon.png',
   scheme: 'expertnabiblia',
@@ -34,7 +34,7 @@ const config = ({ config }: ConfigContext): ExpoConfig => {
   ios: {
     supportsTablet: true,
     bundleIdentifier: 'com.donizetiferr.expertnabiblia',
-    buildNumber: '3',
+    buildNumber: '4',
   },
   android: {
     adaptiveIcon: {
@@ -42,7 +42,7 @@ const config = ({ config }: ConfigContext): ExpoConfig => {
       backgroundColor: '#3c026d',
     },
     package: 'com.donizetiferr.expertnabiblia',
-    versionCode: 3,
+    versionCode: 4,
     permissions: ['NOTIFICATIONS', 'INTERNET'],
   },
   web: {